pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "voxpipe"
description = "Audio pipeline for videos"
authors = [{name = "Alejandro Gaston Alvarez Franceschi", email = "alegasalv@gmail.com"}]
license = "LGPL-3.0"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.11,<3.14"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://alealv.github.io/voxpipe"
Documentation = "https://alealv.github.io/voxpipe"
Changelog = "https://alealv.github.io/voxpipe/changelog"
Repository = "https://github.com/alealv/voxpipe"
Issues = "https://github.com/alealv/voxpipe/issues"
Discussions = "https://github.com/alealv/voxpipe/discussions"
Gitter = "https://gitter.im/voxpipe/community"
Funding = "https://github.com/sponsors/alealv"

[project.scripts]
voxpipe = "voxpipe._internal.cli:app"

[tool.pdm.version]
source = "call"
getter = "scripts.get_version:get_version"

[tool.pdm.build]
# Include as much as possible in the source distribution, to help redistributors.
excludes = ["**/.pytest_cache", "**/.ruff_cache"]
source-includes = [
    "config",
    "docs",
    "scripts",
    "share",
    "tests",
    "duties.py",
    "mkdocs.yml",
    "*.md",
    "LICENSE",
]

[tool.pdm.build.wheel-data]
# Manual pages can be included in the wheel.
# Depending on the installation tool, they will be accessible to users.
# pipx supports it, uv does not yet, see https://github.com/astral-sh/uv/issues/4731.
data = [
    {path = "share/**/*", relative-to = "."},
]

[dependency-groups]
maintain = [
    "build>=1.2",
    "git-changelog>=2.5",
    "twine>=5.1",
    "yore>=0.3.3",
]
ci = [
    "duty>=1.6",
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "ty>=0.0.1a6",
    "deptry>=0.12",
]
 docs = [
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocs>=1.6",
    "mkdocs-coverage>=1.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-llmstxt>=0.2",
    "mkdocs-material>=9.5",
    "mkdocs-minify-plugin>=0.8",
    "mkdocs-section-index>=0.3",
    "mkdocstrings[python]>=0.29",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
]

[tool.uv]
default-groups = ["maintain", "ci", "docs"]

[tool.deptry]
extend_exclude = ["tests", "scripts", "duties.py"]

[tool.deptry.per_rule_ignores]
DEP002 = ["torch"]  # torch is used indirectly via pyannote-audio
